+++ src/timeout-handler/CookieTimeoutHandler.ts
@@ -5,21 +5,21 @@ const GRAVITY_SESSION_TIMEOUT_COOKIE_KEY = 'gravity_session_timeout'
 
 export default class CookieTimeoutHandler implements ITimeoutHandler {
   public constructor(private readonly sessionDuration: number, private readonly win: typeof window) {
-    this.resetTimeout()
+    const stored = readCookie(GRAVITY_SESSION_TIMEOUT_COOKIE_KEY, this.win)
+    if (stored === undefined) {
+      this.reset()
+    }
   }
 
   isExpired(): boolean {
     const stored = readCookie(GRAVITY_SESSION_TIMEOUT_COOKIE_KEY, this.win)
     if (stored === undefined) {
-      this.resetTimeout()
       return false
     }
-    const isExpired = new Date().getTime() >= parseInt(stored)
-    if (isExpired) this.resetTimeout()
-    return isExpired
+    return new Date().getTime() >= parseInt(stored)
   }
 
-  private resetTimeout() {
+  reset() {
     const timeout = new Date().getTime() + this.sessionDuration
     setCookie(GRAVITY_SESSION_TIMEOUT_COOKIE_KEY, timeout.toString(), this.win)
   }